Call C++ calc function from python custom op meta (p2)
- used on conv2d_fp8

+++ hpu_ops/custom_op_outshape.h
@@ -30,7 +30,7 @@ using sizes_vec = std::vector<std::vector<int64_t>>;
       params_int,                            \
       (inputs, params),                      \
       const std::vector<at::Tensor>& inputs, \
-      const std::vector<int>& params)
+      const std::vector<int64_t>& params)
 
 class CustomOpOutShapeFunRegistrar {
  public:
