+++ python_packages/habana_frameworks/torch/hpu/csrc/bindings.cpp
@@ -402,7 +402,7 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
       "custom_op_calc_out_shape_params_int",
       [](const char* opname,
          const std::vector<at::Tensor>& inputs,
-         const std::vector<int>& params) {
+         const std::vector<int64_t>& params) {
         return habana::CustomOpOutShapeFunRegistrar::GetInstance().CalcOutShape(
             opname, inputs, params);
       });
